workspaces/lily_ai_review/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
jinja2>=3.1.2
python-dotenv>=1.0.0
supabase>=1.0.3
python-multipart>=0.0.6
requests>=2.30.0
httpx>=0.24.0
pydantic>=2.0.0
stripe>=5.0.0
asyncio>=3.4.3
